requirements.txt
flask
sqlalchemy
pymysql
python-dotenv
cryptography